make teams and groups work in the cli
Add the fundamental necessities for getting the CLI
callable for teams and groups channel messages.

This is the second in a series of PRs to get v0 backups
working for channels. In this change, the current api
enumerators get plugged into the collections
handler to produce backup data. Follow-up PRs will:

* fix any bugs preventing backups from running via the cli
* swap full-item enumeration for id-first-get-later pattern
* populate each message with all its replies on the get-later
* turn on integration testing at the operations and ci layers